Simplify components with styled-component
Simplify components with styled-component
