+++ .flowconfig
@@ -2,6 +2,9 @@
 ; We fork some components by platform
 .*/*[.]android.js
 
+; Ignore styled components vendor src directroy
+.*/styled-components/src/.*
+
 ; Ignore "BUCK" generated dirs
 <PROJECT_ROOT>/\.buckd/
