+++ js/components/WLHeader.js
@@ -4,8 +4,7 @@
 'use strict';
 
 import React, { Component } from 'react';
-import styles from './style';
-import { View, Text, TouchableOpacity } from 'react-native';
+import { View, Text, TouchableOpacity, Platform, StyleSheet } from 'react-native';
 
 class Header extends  Component {
 
@@ -45,5 +44,41 @@ class Header extends  Component {
     );
   }
 }
+
+const STATUS_BAR_HEIGHT = Platform.OS === 'ios' ? 20 : 25;
+const HEADER_HEIGHT = Platform.OS === 'ios' ? 44 + STATUS_BAR_HEIGHT : 56 + STATUS_BAR_HEIGHT;
+
+const styles = StyleSheet.create({
+  container: {
+    height: HEADER_HEIGHT - STATUS_BAR_HEIGHT,
+    marginBottom: 30,
+  },
+  header: {
+    height: HEADER_HEIGHT,
+    paddingTop: STATUS_BAR_HEIGHT,
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    alignItems: 'center',
+    padding: 10,
+    backgroundColor: '#4395f7',
+  },
+  leftItem: {
+    flex: 1,
+    alignItems: 'flex-start',
+  },
+  centerItem: {
+    flex: 2,
+    alignItems: 'center',
+  },
+  rightItem: {
+    flex: 1,
+    alignItems: 'flex-end',
+  },
+  titleText: {
+    color: '#fff',
+    fontWeight: 'bold',
+    fontSize: 20,
+  },
+});
 
 export default Header;
