+++ js/components/WLPlayer.js
@@ -4,9 +4,8 @@
 'use strict';
 
 import React from 'react';
-import type {Id, Score} from '../../types/';
-import styles from './style';
-import {View, Text, TouchableHighlight} from 'react-native';
+import type {Id, Score} from '../types/';
+import {View, Text, TouchableHighlight, StyleSheet } from 'react-native';
 
 type Props = {
   id: Id,
@@ -62,5 +61,92 @@ class Player extends React.Component {
     );
   }
 }
+
+const styles = StyleSheet.create({
+  wrapper: {
+    flex: 1,
+    borderRadius: 10,
+    backgroundColor: '#4395f7',
+  },
+  name: {
+    position: 'absolute',
+    zIndex: 10,
+    top: 35,
+    left: 0,
+    right: 0,
+  },
+  nameBody: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  nameText: {
+    fontSize: 20,
+    color: '#a3d6f5',
+  },
+  score: {
+    position: 'absolute',
+    zIndex: 20,
+    top: 0,
+    bottom: 0,
+    left: 0,
+    right: 0,
+  },
+  scoreBody: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+  scoreText: {
+    color: '#ffffff',
+    fontSize: 140,
+  },
+  buttons: {
+    position: 'absolute',
+    zIndex: 30,
+    top: 0,
+    bottom: 0,
+    left: 0,
+    right: 0,
+  },
+  buttonsBody: {
+    flex: 1,
+    flexDirection: 'row',
+  },
+  left: {
+    flex: 1,
+    flexDirection: 'row',
+    paddingTop: 10,
+    paddingRight: 10,
+    paddingBottom: 0,
+    paddingLeft: 10,
+    borderTopLeftRadius: 10,
+    borderBottomLeftRadius: 10,
+  },
+  right: {
+    flex: 1,
+    flexDirection: 'row',
+    paddingTop: 10,
+    paddingRight: 10,
+    paddingBottom: 0,
+    paddingLeft: 10,
+    borderTopRightRadius: 10,
+    borderBottomRightRadius: 10,
+  },
+  leftText: {
+    flex: 1,
+    textAlign: 'left',
+    color: '#a3d6f5',
+    fontSize: 50,
+    alignSelf: 'flex-end',
+  },
+  rightText: {
+    flex: 1,
+    textAlign: 'right',
+    color: '#a3d6f5',
+    fontSize: 50,
+    alignSelf: 'flex-end',
+  }
+});
 
 export default Player;
