--- js/components/header/style.js
@@ -1,44 +1,0 @@
-/**
- * @flow
- */
-'use strict';
-
-import { Platform, StyleSheet } from 'react-native';
-
-var STATUS_BAR_HEIGHT = Platform.OS === 'ios' ? 20 : 25;
-var HEADER_HEIGHT = Platform.OS === 'ios' ? 44 + STATUS_BAR_HEIGHT : 56 + STATUS_BAR_HEIGHT;
-
-const styles = StyleSheet.create({
-  container: {
-    height: HEADER_HEIGHT - STATUS_BAR_HEIGHT,
-    marginBottom: 30,
-  },
-  header: {
-    height: HEADER_HEIGHT,
-    paddingTop: STATUS_BAR_HEIGHT,
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    padding: 10,
-    backgroundColor: '#4395f7',
-  },
-  leftItem: {
-    flex: 1,
-    alignItems: 'flex-start',
-  },
-  centerItem: {
-    flex: 2,
-    alignItems: 'center',
-  },
-  rightItem: {
-    flex: 1,
-    alignItems: 'flex-end',
-  },
-  titleText: {
-    color: '#fff',
-    fontWeight: 'bold',
-    fontSize: 20,
-  },
-});
-
-export default styles;
