--- js/components/utility/style.js
@@ -1,19 +1,0 @@
-/**
- * @flow
- */
-'use strict';
-
-import {StyleSheet} from 'react-native';
-const styles = StyleSheet.create({
-  wrapper: {
-    margin: 10,
-    flex: 0,
-    flexDirection: 'row',
-    justifyContent: 'space-around',
-  },
-  text: {
-    color: '#fff',
-  }
-});
-
-export default styles;
