+++ package.json
@@ -13,7 +13,8 @@
     "react": "15.4.1",
     "react-native": "0.39.0",
     "react-redux": "^4.4.6",
-    "redux": "^3.6.0"
+    "redux": "^3.6.0",
+    "styled-components": "^1.2.1"
   },
   "devDependencies": {
     "babel-eslint": "^7.1.1",
